refactor getDirectorySize->directoryHelper
